Deduplicate and simplify request tests
- only a small chunk of tests is expected to yield
  different result on different node type
- clearly mark such cases
- test sensitive parts against both node types
